backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
requests==2.28.1
requests_cache==0.4.12
Web3==7.9.0
certifi==2024.2.2
setuptools
python-dotenv==0.21.1
nillion-client==0.2.0
secretvaults==0.0.0a7
nest_asyncio==1.6.0